requirements.txt
lark-sdk
lancedb
openai
python-dotenv